Give up on PK on observation_fact; replace indexes
Any relaxed PK constraint may be appropriate for transmart, but chokes
on i2b2 demodata. We can revisisit this if we give up on supporting i2b2
data.

The indexes on observation_fact are also not appropriate for transmart,
so they were replaced with a simpler one that's tailored for the way
clinical data queries are done in transmart.